+++ crates/pdk-api/src/extension.rs
@@ -16,10 +16,6 @@ api_struct!(
 api_struct!(
     /// Output returned from the `register_extension` function.
     pub struct RegisterExtensionOutput {
-        /// Schema shape of the tool's configuration.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub config_schema: Option<Schema>,
-
         /// Name of the extension.
         pub name: String,
 
@@ -28,6 +24,14 @@ api_struct!(
 
         /// Version of the plugin.
         pub plugin_version: String,
+    }
+);
+
+api_struct!(
+    /// Output returned from the `define_extension_config` function.
+    pub struct DefineExtensionConfigOutput {
+        /// Schema shape of the extension's configuration.
+        pub schema: Schema,
     }
 );