+++ CHANGELOG.md
@@ -7,13 +7,17 @@
 - **WASM API**
   - Renamed `ProjectFragment.alias` to `ProjectFragment.aliases` and changed its type from
     `Option<String>` to `Vec<String>`.
+  - Removed `RegisterExtensionOutput.config_schema` field. Use the new `define_extension_config`
+    plugin function instead.
 
 #### 🚀 Updates
 
 - **WASM API**
   - Added a `load_extension_config_by_id` host function.
+  - Added a `define_extension_config` plugin function.
   - Added `load_extension_config`, `parse_extension_config` and `parse_extension_config_schema`
     utility functions.
+  - Added `DefineExtensionConfigOutput` type.
   - Added `ExtendProjectGraphInput.extension_config` field.
   - Added `ExtendTaskCommandInput.extension_config` field.
   - Added `ExtendTaskScriptInput.extension_config` field.